revert fix for cycle 2
