+++ lib/eventContext.ts
@@ -1,7 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache License. See License in the project root for license information.
 
-import type { Connection, Container } from "./connection";
+import type { Connection } from "./connection";
+import type { Container } from "./container";
 import { Session } from "./session";
 import {
   Delivery, Message, ConnectionError, EventContext as RheaEventContext
