+++ lib/index.ts
@@ -12,16 +12,9 @@ export {
 } from "rhea";
 
 export { EventContext, OnAmqpEvent } from "./eventContext";
+export { Container, ContainerOptions } from "./container";
 export {
-  Connection,
-  ReqResLink,
-  ConnectionOptions,
-  CreateReceiverOptions,
-  CreateAwaitableSenderOptions,
-  CreateSenderOptions,
-  CreateRequestResponseLinkOptions,
-  Container,
-  ContainerOptions,
+  Connection, ReqResLink, ConnectionOptions, CreateReceiverOptions, CreateAwaitableSenderOptions, CreateSenderOptions, CreateRequestResponseLinkOptions
 } from "./connection";
 export { Session } from "./session";
 export { Receiver, ReceiverOptions } from "./receiver";
